moves react components to app/javascript/bundles/App/components
