+++ app/javascript/bundles/App/components/HiveRow.jsx
