+++ app/javascript/bundles/App/components/HivesTable.jsx
