+++ app/javascript/bundles/App/components/InspectionRow.jsx
