+++ app/javascript/bundles/App/components/InspectionsTable.jsx
