+++ app/javascript/bundles/App/components/MenuBar.jsx
